chore: split pruning.rs into four files.
